[NEW] Backend: Infraestructure fixes

+++ E-commerce web application/Backend/Application/Response/DataBalanceResponse.cs
@@ -4,10 +4,10 @@
     {
         public string? FirstNameClient { get; set; }
         public string? LastNameClient { get; set; }
-        public double? Total { get; set; }
-        public double? SubTotal { get; set; }
+        public decimal? Total { get; set; }
+        public decimal? SubTotal { get; set; }
         public string? ProductName { get; set; }
         public int? ProductAmount { get; set; }
-        public double? ProductPrice { get; set; }
+        public decimal? ProductPrice { get; set; }
     }
 }
