+++ E-commerce web application/Backend/Application/UserCase/Order/OrderProductData.cs
@@ -4,7 +4,7 @@
     {
         public int CartId{ get; set; }
         public int Amount{ get; set; }
-        public double Price{ get; set; }
-        public double Total{ get; set; }
+        public decimal Price{ get; set; }
+        public decimal Total{ get; set; }
     }
 }
