+++ E-commerce web application/Backend/Application/UserCase/Product/ProductData.cs
@@ -3,7 +3,7 @@
     public class ProductData
     {
         public ProductData() { }
-        public ProductData(int id, int cantidad, double precio)
+        public ProductData(int id, int cantidad, decimal precio)
         {
             Id = id;
             Amount = cantidad;
@@ -12,6 +12,6 @@
 
         public int Id { get; set; }
         public int Amount { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
     }
 }
